Moved SolarWinds class into orionsdk folder
Moved the SolarWinds class in the __init__.py file in the samples folder into the orionsdk folder as solarwinds.py.  Updated __init__.py in the orionsdk folder to import the SolarWinds class.  This is related to issue #39.
